src/main/java/com/mycompany/app/App.java
package com.mycompany.app;
import java.util.*;
import spark.ModelAndView;
import spark.template.mustache.MustacheTemplateEngine;
import static spark.Spark.*;
public class App {
	public static boolean append_search(ArrayList<Integer> array1, ArrayList<Integer> array2, int e) {
		ArrayList<Integer> array = new ArrayList<Integer>();
		if(array1 == null && array2 == null)
			return false;
		if(array1 != null) {
			for(int x : array1)
				array.add(x);
		}
		if(array2 != null || !array2.isEmpty()) {
			for(int y : array2)
				array.add(y);
		}
				
		if(array != null || !array.isEmpty()) {
			for(int a : array) {
			if(a==e)
				return true;
			}
		}
		return false;
	}
public static void main(String[] args) {
        port(getHerokuAssignedPort());

        get("/", (req, res) -> "Hello, World");

        post("/compute", (req, res) -> {
          //System.out.println(req.queryParams("input1"));
          //System.out.println(req.queryParams("input2"));
	  //System.out.println(req.queryParams("input3"));

          String input1 = req.queryParams("input1");
          java.util.Scanner sc1 = new java.util.Scanner(input1);
          sc1.useDelimiter("[;\r\n]+");
          java.util.ArrayList<Integer> inputList1 = new java.util.ArrayList<>();
          while (sc1.hasNext())
          {
            int value = Integer.parseInt(sc1.next().replaceAll("\\s",""));
            inputList1.add(value);
          }
          System.out.println(inputList1);

	String input2 = req.queryParams("input2");
          java.util.Scanner sc2 = new java.util.Scanner(input2);
          sc2.useDelimiter("[;\r\n]+");
          java.util.ArrayList<Integer> inputList2 = new java.util.ArrayList<>();
          while (sc2.hasNext())
          {
            int value = Integer.parseInt(sc2.next().replaceAll("\\s",""));
            inputList2.add(value);
          }
          System.out.println(inputList2);


          String input3 = req.queryParams("input3").replaceAll("\\s","");
          int input2AsInt = Integer.parseInt(input3);

          boolean result = App.append_search(inputList1, inputList2, input2AsInt);

         Map map = new HashMap();
          map.put("result", result);
          return new ModelAndView(map, "compute.mustache");
        }, new MustacheTemplateEngine());


        get("/compute",
            (rq, rs) -> {
              Map map = new HashMap();
              map.put("result", "not computed yet!");
              return new ModelAndView(map, "compute.mustache");
            },
	new MustacheTemplateEngine());
    }

    static int getHerokuAssignedPort() {
        ProcessBuilder processBuilder = new ProcessBuilder();
        if (processBuilder.environment().get("PORT") != null) {
            return Integer.parseInt(processBuilder.environment().get("PORT"));
        }
        return 4567; //return default port if heroku-port isn't set (i.e. on localhost)
    }


}
